Moved error report function to new file
  - Will be used by all SUNDIALS functions from different places of C
runtime
